Check services are available before starting the listener
Testerina normally starts the given listeners during the initialization. When we try to explicitly attach and start a service within a test, then the start method calls twice(first by Testerina and then our test). The first call by Testerina is not going to be successful because at that time there is no service attached to the listener. But that first call sets a flag indicating service is started. Due to the flag, the second call will not be triggered. Here, we add a condition into the start method that checks whether a service is attached or not.