Fix the backward incompatibility by handling the InvalidProtocolBufferException.
CodeInputStream.readTag function returns and InvalidProtocolBufferException when it could not find the field in the wire. Before throwing that exception, it updates the last tag with the currently reading tag (unknown tag). Here, we handle the exception and assign the last tag value to the currently executing tag variable. Although we set the variable, it won't process because the unknown tag is not in the protobuf file. But anyway, it returns the expected message to the user without throwing an exception.