Fix invalid check for service names in lower camel case
The common practice is to have gRPC service names in the upper camel case. However, when someone defines service with a lower camel case name, Ballerina gRPC CLI generates callers by creating names in the upper camel case. Even the Ballerina gRPC runtime anticipates a caller name in the upper camel case.

In the compiler plugin, we have a validation to check whether the caller names are starting with the service name or not. That compiler plugin validation is incorrect for the above proto file. The correct behaviour should be checking the caller name against the service name after converting it to the upper camel case.